+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 ## [1.0.0] - 2022-03-02
  - Initial release
+
+## [1.1.0] - 2022-03-02
+ - Store the host, port and database into the session and auto fill it when loading the form.
+ - Store the version into the session and clear it when it is changed.